% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwa.R
\name{fwa_coastline}
\alias{fwa_coastline}
\title{Read coastline polylines from FWA_COASTLINES_SP layer.}
\usage{
fwa_coastline(coastline,
  dsn = "~/Poisson/Data/spatial/fwa/gdb/FWA_BC.gdb")
}
\arguments{
\item{coastline}{A vector of valid coastline WatershedGroupCode, WatershedGroupName, BlueLineKey, and/or WatershedCode. (see fwa_coastline_wsg_lookup and fwa_coastline_blk_lookup for reference).}

\item{dsn}{A character string indicating path to FWA database with FWA_COASTLINE_SP layer.}
}
\value{
A sf object.
}
\description{
Select particular coastlines with coastline argument.
}
\examples{
porcher <- fwa_coastline("Porcher Island")
}
