% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fwa_lookup_layer}
\alias{fwa_lookup_layer}
\title{FWA layer lookup}
\format{A data frame with 11 rows and 3 variables:
\describe{
  \item{layer}{freshwater atlas layer}
  \item{WATERSHED_KEY}{whether it can be filtered by WATERSHED_KEY}
  \item{WATERSHED_GROUP_CODE}{whether it can be filtered by WATERSHED_GROUP_CODE}
  ...
}}
\usage{
fwa_lookup_layer
}
\description{
A dataset of all freshwater atlas layers that can be used in `fwa_read()`.
}
\keyword{datasets}
