% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fwa_lookup_stream_blkey}
\alias{fwa_lookup_stream_blkey}
\title{Stream BlueLineKey lookup table}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\usage{
fwa_lookup_stream_blkey
}
\description{
A dataset of all stream BlueLineKey and WatershedCode.
}
\keyword{datasets}
