% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fwa_lookup_watershed_group}
\alias{fwa_lookup_watershed_group}
\title{WATERSHED_GROUP_CODE lookup table}
\format{A data frame with 18614 rows and 12 variables:
\describe{
  \item{GNIS_NAME}{gnis name}
  \item{WATERSHED_KEY}{fwa watershed key}
  \item{stream-network}{whether data is available for stream-network layer}
  \item{linear-boundaries}{whether data is available for linear-boundaries layer}
  \item{watersheds}{whether data is available for watersheds layer}
  \item{coastlines}{whether data is available for coastlines layer}
  \item{obstructions}{whether data is available for obstructions layer}
  \item{lakes}{whether data is available for lakes layer}
  \item{wetlands}{whether data is available for wetlands layer}
  \item{rivers}{whether data is available for rivers layer}
  \item{manmade-waterbodies}{whether data is available for manmade-waterbodies layer}
  \item{glaciers}{wwhether data is available for glaciers layer}
  ...
}}
\usage{
fwa_lookup_watershed_group
}
\description{
A dataset of all WATERSHED_GROUP_CODES and which layers are available for each.
}
\keyword{datasets}
