% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{fwa_pull_tributaries}
\alias{fwa_pull_tributaries}
\title{Pull tributaries from WATERSHED_KEY.}
\usage{
fwa_pull_tributaries(x, order = 1)
}
\arguments{
\item{x}{A vector of WATERSHED_KEY.}

\item{order}{An integer of the tributary order,
e.g. order = 1L will retrieve all primary upstream tributaries.}
}
\value{
A vector of WATERSHED_KEY tributaries.
}
\description{
Pull tributaries from WATERSHED_KEY.
}
\examples{
\dontrun{
fwa_pull_tributaries(360709847)
}
}
