% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{fwa_pull_watershed_group_code}
\alias{fwa_pull_watershed_group_code}
\title{Pull WATERSHED_GROUP_CODE from WATERSHED_GROUP_NAME.}
\usage{
fwa_pull_watershed_group_code(x)
}
\arguments{
\item{x}{A vector of valid WATERSHED_GROUP_NAME.}
}
\value{
A character vector.
}
\description{
Pull WATERSHED_GROUP_CODE from WATERSHED_GROUP_NAME.
}
\examples{
\dontrun{
fwa_pull_watershed_group_code("Graham Island")
}
}
