% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{fwa_pull_watershed_key}
\alias{fwa_pull_watershed_key}
\title{Pull WATERSHED_KEY from GNIS_NAME.}
\usage{
fwa_pull_watershed_key(x)
}
\arguments{
\item{x}{A vector of valid GNIS_NAME.}
}
\value{
A character vector.
}
\description{
Pull WATERSHED_KEY from GNIS_NAME.
}
\examples{
fwa_pull_watershed_key("Sangan River")
}
