% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_coastline}
\alias{fwa_read_coastline}
\title{Read coastline from FWA_COASTLINES_SP layer in FWA_BC geodatabase.}
\usage{
fwa_read_coastline(x = NULL, ask = FALSE)
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY or WATERSHED_GROUP_CODE.}

\item{ask}{A flag indicating whether to ask before reading entire dataset.}
}
\value{
A sf object.
}
\description{
Read coastline from FWA_COASTLINES_SP layer in FWA_BC geodatabase.
}
\examples{
coastline <- fwa_read_coastline("GRAI")
}
