% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_coastlines}
\alias{fwa_read_coastlines}
\title{Read from coastlines layer.}
\usage{
fwa_read_coastlines(x = NULL, ask = FALSE, crs = 3005)
}
\value{
A sf object with sfc_LINESTRING geometry.
}
\description{
Read from coastlines layer.
}
\examples{
\dontrun{
fwa_read_coastlines("GRAI")
}
}
