% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_lakes}
\alias{fwa_read_lakes}
\title{Read from lakes layer.}
\usage{
fwa_read_lakes(x = NULL, ask = FALSE, crs = 3005)
}
\value{
A sf object with sfc_POLYGON geometry.
}
\description{
Read from lakes layer.
}
\examples{
\dontrun{
fwa_read_lakes("GRAI")
}
}
