% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_layer}
\alias{fwa_read_layer}
\title{Read features from FWA database layer.}
\usage{
fwa_read_layer(x = NULL, layer = "stream-network", crs = 3005,
  ask = TRUE)
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY or WATERSHED_GROUP_CODE. If NULL, entire dataset is read.}

\item{layer}{A character string of valid layer name. See fwa_lookup_layer.}

\item{crs}{The epsg code for the coordinate reference system. Defaults to `3005`
(B.C. Albers). See https://epsgi.io.}

\item{ask}{A flag indicating whether to ask before reading entire dataset.}
}
\value{
A sf object.
}
\description{
Read features from FWA database layer.
}
\examples{
\dontrun{
fwa_read_layer(c("VICT", 360843586), layer = "stream-network")
}
}
