% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_layers}
\alias{fwa_read_layers}
\title{Read features from FWA database layers.}
\usage{
fwa_read_layers(x = NULL, ask = TRUE, layers = fwa_lookup_layer,
  crs = 3005, remove_empty = FALSE)
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY or WATERSHED_GROUP_CODE. If NULL, entire dataset is read.}

\item{ask}{A flag indicating whether to ask before reading entire dataset.}

\item{layers}{A vector of valid character string layer names. See fwa_lookup_layer.}

\item{crs}{The epsg code for the coordinate reference system. Defaults to `3005`
(B.C. Albers). See https://epsgi.io.}

\item{remove_empty}{A flag indicating whether to remove list elements of layers with no data.}
}
\value{
A named list of sf objects.
If no data available and remove_empty = FALSE, value of list element will be NULL.
}
\description{
Read features from FWA database layers.
}
\examples{
\dontrun{
fwa_read_layers(c("VICT"), layers = c("stream-network", "watersheds"), remove_empty = TRUE)
}
}
