% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_linear_boundaries}
\alias{fwa_read_linear_boundaries}
\title{Read from linear-boundaries layer.}
\usage{
fwa_read_linear_boundaries(x = NULL, ask = FALSE, crs = 3005)
}
\value{
A sf object with sfc_LINESTRING geometry.
}
\description{
Read from linear-boundaries layer.
}
\examples{
\dontrun{
fwa_read_linear_boundaries("GRAI")
}
}
