% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_manmade_waterbodies}
\alias{fwa_read_manmade_waterbodies}
\title{Read from manmade-waterbodies layer.}
\usage{
fwa_read_manmade_waterbodies(x = NULL, ask = FALSE, crs = 3005)
}
\value{
A sf object.
}
\description{
Read from manmade-waterbodies layer.
}
\examples{
\dontrun{
fwa_read_manmade_waterbodies("GRAI")
}
}
