% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_obstructions}
\alias{fwa_read_obstructions}
\title{Read from obstructions layer.}
\usage{
fwa_read_obstructions(x = NULL, ask = FALSE, crs = 3005)
}
\value{
A sf object with sfc_POINT geometry.
}
\description{
Read from obstructions layer.
}
\examples{
\dontrun{
fwa_read_obstructions("GRAI")
}
}
