% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_rivers}
\alias{fwa_read_rivers}
\title{Read from rivers layer.}
\usage{
fwa_read_rivers(x = NULL, ask = FALSE, crs = 3005)
}
\value{
A sf object with sfc_POLYGON geometry.
}
\description{
Read from rivers layer.
}
\examples{
\dontrun{
fwa_read_rivers("GRAI")
}
}
