% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_stream}
\alias{fwa_read_stream}
\title{Read features from stream-network layer.}
\usage{
fwa_read_stream(x = NULL, ask = TRUE)
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY or WATERSHED_GROUP_CODE. If NULL, entire dataset is read.}

\item{ask}{A flag indicating whether to ask before reading entire dataset.}
}
\value{
A sf object.
}
\description{
Read features from stream-network layer.
}
\examples{
all <- fwa_read_stream()
streams <- fwa_read_stream(c(360709847, 360843586))
}
