% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_stream_network}
\alias{fwa_read_stream_network}
\title{Read from stream-network layer.}
\usage{
fwa_read_stream_network(x = NULL, ask = TRUE, crs = 3005)
}
\value{
A sf object with sfc_LINESTRING geometry.
}
\description{
Read from stream-network layer.
}
\examples{
\dontrun{
fwa_read_stream_network(ask = FALSE)
fwa_read_stream_network(c(360709847, 360843586))
}
}
