% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_watershed_group}
\alias{fwa_read_watershed_group}
\title{Read watershed group from FWA_WATERSHED_GROUPS layer in FWA_BC geodatabase}
\usage{
fwa_read_watershed_group(x = NULL, ask = FALSE)
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY or WATERSHED_GROUP_CODE.}

\item{ask}{A flag indicating whether to ask before reading entire dataset.}
}
\value{
A sf object.
}
\description{
Read watershed group from FWA_WATERSHED_GROUPS layer in FWA_BC geodatabase
}
\examples{
graham <- fwa_read_watershed_group("GRAI")
}
