% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-layer.R
\name{fwa_read_watershed_groups}
\alias{fwa_read_watershed_groups}
\title{Read from watershed-groups layer.}
\usage{
fwa_read_watershed_groups(x = NULL, crs = 3005, collect = TRUE,
  check = TRUE)
}
\arguments{
\item{x}{A vector of valid WATERSHED_GROUP_NAME or WATERSHED_GROUP_CODE.
If NULL, entire dataset is read.}
}
\value{
A sf object with sfc_POLYGON geometry.
}
\description{
Read from watershed-groups layer.
}
\examples{
\dontrun{
fwa_read_watershed_groups("GRAI")
}
}
