% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_watersheds}
\alias{fwa_read_watersheds}
\title{Read from watersheds layer.}
\usage{
fwa_read_watersheds(x = NULL, ask = TRUE, crs = 3005)
}
\value{
A sf object with sfc_POLYGON geometry.
}
\description{
Read from watersheds layer.
}
\examples{
\dontrun{
fwa_read_watersheds(c(360709847, 360843586))
}
}
