% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_wetlands}
\alias{fwa_read_wetlands}
\title{Read from wetlands layer.}
\usage{
fwa_read_wetlands(x = NULL, ask = FALSE, crs = 3005)
}
\value{
A sf object with sfc_POLYGON geometry.
}
\description{
Read from wetlands layer.
}
\examples{
\dontrun{
fwa_read_wetlands("GRAI")
}
}
