% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{fwa_rkm}
\alias{fwa_rkm}
\title{Generate and lable Rkm points at specified distance a line.}
\usage{
fwa_rkm(x = 360709847, distance = 10, label_name = "Rkm",
  stream_name = "WATERSHED_KEY", sfc_name = "geometry")
}
\arguments{
\item{x}{A valid WATERSHED_KEY.}

\item{distance}{A number indicating distance in metres between each point.}

\item{label_name}{A character string of the name of the new column containing km labels.}

\item{sfc_name}{A character string of the name of the new sfc column containing geometries.}

\item{blkey_name}{A character string of the name of the new column containing BlueLineKey.}
}
\value{
A sf object.
}
\description{
Generate and lable Rkm points at specified distance a line.
}
\examples{
rkm <- fwa_rkm(360709847, distance = 10)
}
