% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{fwa_search_gnis}
\alias{fwa_search_gnis}
\title{Search valid gnis names.}
\usage{
fwa_search_gnis(pattern, ignore_case = TRUE, ...)
}
\arguments{
\item{pattern}{A character string containing a regular expression (or character string for fixed = TRUE) to be matched.}

\item{ignore_case}{A flag indicating whether to ignore case.}

\item{...}{Arguments passed to grepl.}
}
\value{
A character vector.
}
\description{
Search valid gnis names.
}
\examples{
fwa_search_gnis("kaslo")
}
