% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{fwa_search_gnis_name}
\alias{fwa_search_gnis_name}
\title{Search valid GNIS_NAME.}
\usage{
fwa_search_gnis_name(pattern, ignore_case = TRUE, ...)
}
\arguments{
\item{pattern}{A character string containing a regular expression (or character string for fixed = TRUE) to be matched.}

\item{ignore_case}{A flag indicating whether to ignore case.}

\item{...}{Arguments passed to grepl.}
}
\value{
A character vector.
}
\description{
Search valid GNIS_NAME.
}
\examples{
fwa_search_gnis_name("sangan")
}
