% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{fwa_search_wsgroup}
\alias{fwa_search_wsgroup}
\title{Search valid watershed groups.}
\usage{
fwa_search_wsgroup(pattern, ignore_case = TRUE, coast_only = FALSE,
  code = FALSE, ...)
}
\arguments{
\item{pattern}{A character string containing a regular expression (or character string for fixed = TRUE) to be matched.}

\item{ignore_case}{A flag indicating whether to ignore case.}

\item{coast_only}{A flag indicating whether to only search watershed groups in the coastline layer.}

\item{code}{A flag indicating whether to search by watershed group code (not name).}

\item{...}{Arguments passed to grepl.}
}
\value{
A character vector.
}
\description{
Search valid watershed groups.
}
\examples{
fwa_search_wsgroup("graham")
}
