% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fwa_stream_lookup}
\alias{fwa_stream_lookup}
\title{Stream lookup table}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\usage{
fwa_stream_lookup
}
\description{
A dataset of all streams in useful coding systems.
}
\keyword{datasets}
