% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwa.R
\name{fwa_watershed}
\alias{fwa_watershed}
\title{Read watershed polygons from FWA_WATERSHEDS_POLY database layers.}
\usage{
fwa_watershed(x = NULL, tributaries = FALSE,
  dsn = "~/Poisson/Data/spatial/fwa/gdb/FWA_WATERSHEDS_POLY.gdb")
}
\arguments{
\item{x}{A vector of valid GNIS_NAME or FWA_WATERSHED_CODE from FWA_WATERSHEDS_POLY geodatabase}

\item{tributaries}{A flag indicating whether to include tributaries or an integer indicating order of tributaries to include. If TRUE, all tributaries are returned.}

\item{dsn}{A character string indicating path to FWA database with FWA_ROUTES_SP layer.}
}
\value{
A polygon sf object.
}
\description{
Read watershed polygons from FWA_WATERSHEDS_POLY database layers.
}
\examples{
all_grai <- fwa_watershed("Graham Island")
wsheds <- fwa_watershed(c("Hiellen River", "Sangan River"), tributaries = 1L)
}
