% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwa.R
\name{fwa_watershed_group}
\alias{fwa_watershed_group}
\title{Read watershed group polygons from FWA_WATERSHED_GROUPS_POLY layer.}
\usage{
fwa_watershed_group(watershed_group = "Graham Island", dsn)
}
\arguments{
\item{watershed_group}{A vector of valid WatershedGroupCode and/or WatershedGroupNames (see fwa_wsgroup_lookup reference).}

\item{dsn}{A character string indicating path to FWA database with FWA_ROUTES_SP layer.}
}
\value{
A polygon sf object.
}
\description{
Select particular streams with stream argument. Add all tributaries with tributaries = TRUE
}
\examples{
grai <- fwa_watershed_group("Graham Island")
}
