% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_gk}
\alias{read_gk}
\title{Read layer from GNIS_NAME or WATERSHED_KEY}
\usage{
read_gk(...)
}
\arguments{
\item{x}{A vector of valid GNIS_NAME or WATERSHED_KEY.
If NULL, entire dataset is read.}
}
\description{
Read layer from GNIS_NAME or WATERSHED_KEY
}
