% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_kcn}
\alias{read_kcn}
\title{Read layer from WATERSHED_KEY, WATERSHED_GROUP_NAME or WATERSHED_GROUP_CODE}
\usage{
read_kcn(...)
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY, WATERSHED_GROUP_NAME or WATERSHED_GROUP_CODE.
If NULL, entire dataset is read.}
}
\description{
Read layer from WATERSHED_KEY, WATERSHED_GROUP_NAME or WATERSHED_GROUP_CODE
}
