% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read}
\alias{fwa_read}
\title{Read layer from FWA database.}
\usage{
fwa_read(x = NULL, ask = TRUE, layer = "stream-network")
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY or WATERSHED_GROUP_CODE. If NULL, entire dataset is read.}

\item{ask}{A flag indicating whether to ask before reading entire dataset.}

\item{layer}{A character string of valid layer name (see fwa_lookup_layers for reference).}
}
\value{
A sf object.
}
\description{
Read layer from FWA database.
}
\examples{
streams <- fwa_read_layer(c(360709847, 360843586), layer = "stream-network")
}
