% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-layer.R
\name{fwa_read_coastlines}
\alias{fwa_read_coastlines}
\title{Read from coastlines layer.}
\usage{
fwa_read_coastlines(x = NULL, crs = 3005, collect = TRUE,
  check = TRUE, input_type = NULL)
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY, WATERSHED_GROUP_NAME or WATERSHED_GROUP_CODE.
If NULL, entire dataset is read.}
}
\value{
A sf object with sfc_LINESTRING geometry.
}
\description{
Read from coastlines layer.
}
\examples{
\dontrun{
fwa_read_coastlines("GRAI")
}
}
