% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_lakes}
\alias{fwa_read_lakes}
\title{Read from lakes layer.}
\usage{
fwa_read_lakes(x = NULL, ask = FALSE, crs = 3005)
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY or WATERSHED_GROUP_CODE. If NULL, entire dataset is read.}

\item{ask}{A flag indicating whether to ask before reading entire dataset.}

\item{crs}{The epsg code for the coordinate reference system. Defaults to `3005`
(B.C. Albers). See https://epsgi.io.}
}
\value{
A sf object with sfc_POLYGON geometry.
}
\description{
Read from lakes layer.
}
\examples{
fwa_read_lakes("GRAI")
}
