% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_linear_boundaries}
\alias{fwa_read_linear_boundaries}
\title{Read from linear-boundaries layer.}
\usage{
fwa_read_linear_boundaries(x = NULL, crs = 3005, collect = TRUE,
  check = TRUE)
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY, WATERSHED_GROUP_NAME or WATERSHED_GROUP_CODE.
If NULL, entire dataset is read.}
}
\value{
A sf object with sfc_POLYGON geometry.
}
\description{
Read from linear-boundaries layer.
}
\examples{
\dontrun{
fwa_read_linear_boundaries("GRAI")
}
}
