% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-layer.R
\name{fwa_read_named_watersheds}
\alias{fwa_read_named_watersheds}
\title{Read from named-watersheds layer.}
\usage{
fwa_read_named_watersheds(
  x = NULL,
  tributaries = FALSE,
  named_only = FALSE,
  min_stream_order = 1L,
  crs = 3005,
  collect = TRUE,
  check = TRUE,
  input_type = NULL
)
}
\arguments{
\item{x}{A vector of valid GNIS_NAME or WATERSHED_KEY.
If NULL, entire dataset is read.}

\item{min_stream_order}{An integer indicating minimum STREAM_ORDER to read.}
}
\value{
A sf object with sfc_POLYGON geometry.
}
\description{
Read from named-watersheds layer.
}
\examples{
\dontrun{
fwa_read_named_watersheds("GRAI")
}
}
