% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-network.R
\name{fwa_read_stream_network}
\alias{fwa_read_stream_network}
\title{Read features from 'freshwater-atlas-stream-network' dataset.}
\usage{
fwa_read_stream_network(x = NULL, named_only = FALSE,
  tributaries = FALSE, min_stream_order = 1L, crs = 3005,
  collect = TRUE)
}
\arguments{
\item{x}{A vector of valid GNIS_NAME, WATERSHED_KEY, WATERSHED_GROUP_NAME or WATERSHED_GROUP_CODE.
If NULL, entire dataset is read.}

\item{named_only}{A flag indicating whether to only include features with GNIS_NAME.}

\item{tributaries}{A flag indicating whether to read all tributaries in addition
(only applicable if x is GNIS_NAME or WATERSHED_KEY).}

\item{min_stream_order}{An integer indicating minimum STREAM_ORDER to read.}

\item{crs}{The epsg code for the coordinate reference system. Defaults to `3005`
(B.C. Albers). See https://epsgi.io.}

\item{collect}{A flag indicating whether to collect result.}
}
\value{
A sf object if collect = TRUE, otherwise a object of class bcdc_promise.
}
\description{
Read features from 'freshwater-atlas-stream-network' dataset.
}
\examples{
\dontrun{
fwa_read_stream_network("Fraser River")
}
}
