% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-layer.R
\name{fwa_read_stream_network}
\alias{fwa_read_stream_network}
\title{Read from stream-network layer.}
\usage{
fwa_read_stream_network(
  x = NULL,
  tributaries = FALSE,
  named_only = FALSE,
  min_stream_order = 1L,
  crs = 3005,
  collect = TRUE,
  check = TRUE,
  input_type = NULL
)
}
\arguments{
\item{x}{A vector of valid GNIS_NAME, WATERSHED_KEY, WATERSHED_GROUP_NAME or WATERSHED_GROUP_CODE.
If NULL, entire dataset is read.}

\item{tributaries}{A flag indicating whether to read all tributaries in addition
(only applicable if x is GNIS_NAME or WATERSHED_KEY).}

\item{named_only}{A flag indicating whether to only include features with GNIS_NAME.}

\item{min_stream_order}{An integer indicating minimum STREAM_ORDER to read.}

\item{crs}{The epsg code for the coordinate reference system. Defaults to `3005`
(B.C. Albers). See https://epsgi.io.}

\item{collect}{A flag indicating whether to collect result.}

\item{check}{A flag indicating whether to check that x is valid.}

\item{input_type}{A character string of the column name to filter x by.
If NULL, input_type is guessed. This is useful if input is
both GNIS_NAME and WATERSHED_GROUP_NAME, e.g. 'Skagit River'.}
}
\value{
A sf object with sfc_LINESTRING geometry.
}
\description{
Read from stream-network layer.
}
\examples{
\dontrun{
fwa_read_stream_network("GRAI")
}
}
