% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_stream_network}
\alias{fwa_read_stream_network}
\title{Read from stream-network layer.}
\usage{
fwa_read_stream_network(x = NULL, ask = TRUE, crs = 3005)
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY or WATERSHED_GROUP_CODE. If NULL, entire dataset is read.}

\item{ask}{A flag indicating whether to ask before reading entire dataset.}

\item{crs}{The epsg code for the coordinate reference system. Defaults to `3005`
(B.C. Albers). See https://epsgi.io.}
}
\value{
A sf object with sfc_LINESTRING geometry.
}
\description{
Read from stream-network layer.
}
\examples{
\dontrun{
fwa_read_stream_network(ask = FALSE)
fwa_read_stream_network(c(360709847, 360843586))
}
}
