% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_watershed_groups}
\alias{fwa_read_watershed_groups}
\title{Read from watershed-groups layer.}
\usage{
fwa_read_watershed_groups(x = NULL, ask = FALSE)
}
\arguments{
\item{x}{A vector of valid WATERSHED_GROUP_CODE. If NULL, entire dataset is read.}

\item{ask}{A flag indicating whether to ask before reading entire dataset.}
}
\value{
A sf object.
}
\description{
Read from watershed-groups layer.
}
\examples{
graham <- fwa_read_watershed_group("GRAI")
}
