% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{fwa_read_watersheds}
\alias{fwa_read_watersheds}
\title{Read from watersheds layer.}
\usage{
fwa_read_watersheds(x = NULL, ask = TRUE)
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY or WATERSHED_GROUP_CODE. If NULL, entire dataset is read.}

\item{ask}{A flag indicating whether to ask before reading entire dataset.}
}
\value{
A sf object.
}
\description{
Read from watersheds layer.
}
\examples{
watershed <- fwa_read_watershed(c(360709847, 360843586))
}
