% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watersheds.R
\name{fwa_read_watersheds}
\alias{fwa_read_watersheds}
\title{Read features from 'freshwater-atlas-watersheds' dataset.}
\usage{
fwa_read_watersheds(x = NULL, tributaries = FALSE,
  min_watershed_order = 1L, crs = 3005, collect = TRUE)
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY, WATERSHED_GROUP_NAME or WATERSHED_GROUP_CODE.
If NULL, entire dataset is read.}

\item{tributaries}{A flag indicating whether to read all tributaries in addition
(only applicable if x is GNIS_NAME or WATERSHED_KEY).}

\item{min_watershed_order}{An integer indicating minimum WATERSHED_ORDER to read.}

\item{crs}{The epsg code for the coordinate reference system. Defaults to `3005`
(B.C. Albers). See https://epsgi.io.}

\item{collect}{A flag indicating whether to collect result.}
}
\value{
A sf object if collect = TRUE, otherwise a object of class bcdc_promise.
}
\description{
Read features from 'freshwater-atlas-watersheds' dataset.
}
\examples{
\dontrun{
fwa_read_watersheds("Upper Fraser River")
}
}
