% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-layer.R
\name{fwa_read_watersheds}
\alias{fwa_read_watersheds}
\title{Read from watersheds layer.}
\usage{
fwa_read_watersheds(x = NULL, tributaries = FALSE, crs = 3005,
  collect = TRUE, check = TRUE, input_type = NULL)
}
\arguments{
\item{x}{A vector of valid WATERSHED_KEY, WATERSHED_GROUP_NAME or WATERSHED_GROUP_CODE.
If NULL, entire dataset is read.}
}
\value{
A sf object with sfc_POLYGON geometry.
}
\description{
Read from watersheds layer.
}
\examples{
\dontrun{
fwa_read_watersheds("GRAI")
}
}
