% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{fwa_search_gnis}
\alias{fwa_search_gnis}
\title{Search valid GNIS_NAME.}
\usage{
fwa_search_gnis(pattern, layer = NULL, ...)
}
\arguments{
\item{pattern}{A character string containing a regular expression (or character string for fixed = TRUE) to be matched.}

\item{layer}{A valid layer from fwa_lookup_layer.}

\item{...}{Arguments passed to grepl.}
}
\value{
A character vector.
}
\description{
Search valid GNIS_NAME.
}
\examples{
fwa_search_gnis_name("sangan")
}
