% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwa.R
\name{fwa_stream}
\alias{fwa_stream}
\title{Read stream polylines from FWA_ROUTES_SP layer.}
\usage{
fwa_stream(stream = "Kaslo River", tributaries = FALSE,
  dsn = "~/Poisson/Data/spatial/fwa/gdb/FWA_BC.gdb")
}
\arguments{
\item{stream}{A vector of valid BlueLineKey/GnisName/WatershedCode in any combination (see fwa_stream_lookup reference).}

\item{tributaries}{A flag indicating whether to include all (TRUE) or no (FALSE) tributaries.}

\item{dsn}{A character string indicating path to FWA database with FWA_ROUTES_SP layer.}
}
\value{
A sf object.
}
\description{
Select particular streams with stream argument. Add all tributaries with tributaries = TRUE
}
\examples{
kaslo_tribs <- fwa_stream("Kaslo River", tributaries = TRUE)
}
