% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwa.R
\name{fwa_stream}
\alias{fwa_stream}
\title{Read stream polylines from FWA_ROUTES_SP layer.}
\usage{
fwa_stream(stream = "Kaslo River", tributaries = FALSE,
  dsn = "~/Poisson/Data/spatial/fwa/gdb/FWA_BC.gdb")
}
\arguments{
\item{stream}{A vector of valid GnisName, BlueLineKey or WatershedCode (see fwa_lookup_stream_gnis and fwa_lookup_stream_blkey reference).}

\item{tributaries}{A flag indicating whether to include tributaries.}

\item{dsn}{A character string indicating path to FWA database with FWA_ROUTES_SP layer.}
}
\value{
A linestring sf object.
}
\description{
Read stream polylines from FWA_ROUTES_SP layer.
}
\examples{
streams <- fwa_stream(c("Sangan River", "Hiellen River"))
sangan_tribs <- fwa_stream("Sangan River", tributaries = TRUE)
}
