% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwa.R
\name{fwa_watershed}
\alias{fwa_watershed}
\title{Read watershed polygons from FWA_WATERSHEDS_POLY database layers.}
\usage{
fwa_watershed(watershed = "Kaslo River", tributaries = FALSE,
  union = FALSE,
  dsn = "~/Poisson/Data/spatial/fwa/gdb/FWA_WATERSHEDS_POLY.gdb")
}
\arguments{
\item{watershed}{A vector of valid GnisName, or WatershedCode (see fwa_lookup_stream reference).}

\item{tributaries}{A flag indicating whether to include all (TRUE) or no (FALSE) tributaries.}

\item{union}{A flag indicating whether to union features into a single feature.}

\item{dsn}{A character string indicating path to FWA database with FWA_ROUTES_SP layer.}
}
\value{
A polygon sf object.
}
\description{
Read watershed polygons from FWA_WATERSHEDS_POLY database layers.
}
\examples{
kaslo <- fwa_watershed("Kaslo River", tributaries = TRUE)
}
