% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwa.R
\name{fwa_watershed}
\alias{fwa_watershed}
\title{Read watershed polygons from FWA_WATERSHEDS_POLY database layers.}
\usage{
fwa_watershed(watershed = "Chown Brook",
  watershed_group = "Graham Island", tributaries = FALSE,
  dsn = "~/Poisson/Data/spatial/fwa/gdb/FWA_WATERSHEDS_POLY.gdb")
}
\arguments{
\item{watershed}{A vector of valid GnisName, or WatershedCode within a single WatershedGroup (see fwa_lookup_stream_gnis for reference).}

\item{watershed_group}{A character string of the WatershedGroupName or WatershedGroupCode containing watershed(s).}

\item{tributaries}{A flag indicating whether to include tributaries.}

\item{dsn}{A character string indicating path to FWA_WATERSHEDS_POLY geodatabase.}
}
\value{
A polygon sf object.
}
\description{
Read watershed polygons from FWA_WATERSHEDS_POLY database layers.
}
\examples{
wsheds <- fwa_watershed(c("Hiellen River", "Sangan River"), "Graham Island", tributaries = TRUE)
}
